+++ test/specs/chrome.js
@@ -1,5 +1,8 @@
 import { describe, it } from '../utils/mochaw'
-import { welcomeScenarions } from './scenarios/welcome.js'
+import { welcomeScenarios } from './scenarios/welcome'
+import { documentSelectorScenarios } from './scenarios/documentSelector'
+import { documentScenarios } from './scenarios/document'
+import { faceScenarios } from './scenarios/face'
 const supportedLanguage = ["en", "es"]
 
 const options = {
@@ -12,208 +15,10 @@ describe('Happy Paths', options, ({driver, pageObjects}) => {
   const {documentSelector, welcome, documentUpload, documentUploadConfirmation, verificationComplete, crossDeviceIntro, crossDeviceLink, crossDeviceMobileNotificationSent, crossDeviceMobileConnected, crossDeviceClientSuccess, crossDeviceSubmit, poaIntro, poaDocumentSelection, poaGuidance, common, cameraPermissions, livenessIntro} = pageObjects
 
   supportedLanguage.forEach((lang) => {
-    welcomeScenarions(driver, welcome, localhostUrl, lang)
-  })
-
-  describe('document selection screen', () => {
-
-    supportedLanguage.forEach( (lang) => {
-
-      it('should verify UI elements on the document selection screen', async () => {
-        driver.get(localhostUrl + `?language=${lang}`)
-        const documentSelectorCopy = documentSelector.copy(lang)
-        welcome.primaryBtn.click()
-        documentSelector.verifyTitle(documentSelectorCopy)
-        documentSelector.verifySubtitle(documentSelectorCopy)
-        documentSelector.verifyLabels(documentSelectorCopy)
-        documentSelector.verifyHints(documentSelectorCopy)
-        documentSelector.verifyIcons(documentSelectorCopy)
-      })
-    })
-  })
-
-  describe('DOCUMENT UPLOAD TESTS', () => {
-
-    const goToPassportUploadScreen = async (parameter='') => {
-      driver.get(localhostUrl + parameter)
-      welcome.primaryBtn.click()
-      documentSelector.passportIcon.click()
-    }
-
-    const uploadFileAndClickConfirmButton = async (fileName) => {
-      documentUpload.getUploadInput()
-      documentUpload.upload(fileName)
-      documentUploadConfirmation.confirmBtn.click()
-    }
-
-    supportedLanguage.forEach( (lang) => {
-
-      const documentUploadCopy = documentUpload.copy(lang)
-      const documentUploadConfirmationCopy = documentUploadConfirmation.copy(lang)
-      const verificationCompleteCopy = verificationComplete.copy(lang)
-
-      it('should display cross device UI elements on doc upload screen', async () => {
-        goToPassportUploadScreen(`?language=${lang}`)
-        documentUpload.verifyCrossDeviceUIElements(documentUploadCopy)
-      })
-
-      it('should display uploader icon and button', async () => {
-        goToPassportUploadScreen(`?language=${lang}`)
-        documentUpload.verifyUploaderIcon(documentUploadCopy)
-        documentUpload.verifyUploaderButton(documentUploadCopy)
-      })
-
-      it('should upload a passport and verify UI elements', async () => {
-        goToPassportUploadScreen(`?language=${lang}`)
-
-        documentUpload.verifyPassportTitle(documentUploadCopy)
-        documentUpload.verifyPassportInstructionMessage(documentUploadCopy)
-        documentUpload.getUploadInput()
-        documentUpload.upload('passport.jpg')
-        documentUploadConfirmation.verifyCheckReadabilityMessage(documentUploadConfirmationCopy)
-        documentUploadConfirmation.verifyMakeSurePassportMessage(documentUploadConfirmationCopy)
-      })
-
-      it('should upload driving licence and verify UI elements', async () => {
-        driver.get(localhostUrl + `?language=${lang}`)
-        welcome.primaryBtn.click()
-        documentSelector.drivingLicenceIcon.click()
-        documentUpload.verifyFrontOfDrivingLicenceTitle(documentUploadCopy)
-        documentUpload.verifyFrontOfDrivingLicenceInstructionMessage(documentUploadCopy)
-        documentUpload.getUploadInput()
-        documentUpload.upload('uk_driving_licence.png')
-        documentUploadConfirmation.verifyCheckReadabilityMessage(documentUploadConfirmationCopy)
-        documentUploadConfirmation.verifyMakeSureDrivingLicenceMessage(documentUploadConfirmationCopy)
-        documentUploadConfirmation.confirmBtn.click()
-        documentUpload.verifyBackOfDrivingLicenceTitle(documentUploadCopy)
-        documentUpload.verifyBackOfDrivingLicenceInstructionMessage(documentUploadCopy)
-        documentUpload.getUploadInput()
-        documentUpload.upload('back_driving_licence.jpg')
-        documentUploadConfirmation.verifyCheckReadabilityMessage(documentUploadConfirmationCopy)
-        documentUploadConfirmation.verifyMakeSureDrivingLicenceMessage(documentUploadConfirmationCopy)
-      })
-
-      it('should upload identity card and verify UI elements', async () => {
-        driver.get(localhostUrl + `?language=${lang}`)
-        welcome.primaryBtn.click()
-        documentSelector.identityCardIcon.click()
-        documentUpload.verifyFrontOfIdentityCardTitle(documentUploadCopy)
-        documentUpload.verifyFrontOfIdentityCardInstructionMessage(documentUploadCopy)
-        uploadFileAndClickConfirmButton('national_identity_card.jpg')
-        documentUpload.verifyBackOfIdentityCardTitle(documentUploadCopy)
-        documentUpload.verifyBackOfIdentityCardInstructionMessage(documentUploadCopy)
-        documentUpload.getUploadInput()
-        documentUpload.upload('back_national_identity_card.jpg')
-        documentUploadConfirmation.verifyCheckReadabilityMessage(documentUploadConfirmationCopy)
-        documentUploadConfirmation.verifyMakeSureIdentityCardMessage(documentUploadConfirmationCopy)
-      })
-
-      it('should return no document message after uploading non-doc image', async () => {
-        goToPassportUploadScreen(`?language=${lang}`)
-        uploadFileAndClickConfirmButton('llama.pdf')
-        documentUploadConfirmation.verifyNoDocumentError(documentUploadConfirmationCopy)
-      })
-
-      it('should upload a document on retry', async () => {
-        goToPassportUploadScreen(`?language=${lang}`)
-        uploadFileAndClickConfirmButton('llama.pdf')
-        documentUploadConfirmation.redoBtn.click()
-        documentUpload.getUploadInput()
-        documentUpload.upload('passport.jpg')
-        documentUploadConfirmation.verifyCheckReadabilityMessage(documentUploadConfirmationCopy)
-      })
-
-      it('should return file size too large message for doc', async () => {
-        goToPassportUploadScreen(`?language=${lang}`)
-        documentUpload.getUploadInput()
-        documentUpload.upload('over_10mb_face.jpg')
-        documentUploadConfirmation.verifyFileSizeTooLargeError(documentUploadConfirmationCopy)
-      })
-
-      it('should return use another file type message', async () => {
-        goToPassportUploadScreen(`?language=${lang}`)
-        documentUpload.getUploadInput()
-        documentUpload.upload('unsupported_file_type.txt')
-        documentUploadConfirmation.verifyUseAnotherFileError(documentUploadConfirmationCopy)
-      })
-
-      it('should return unsupported file type error for selfie', async () => {
-        goToPassportUploadScreen(`?language=${lang}&async=false&useWebcam=false`)
-        uploadFileAndClickConfirmButton('passport.jpg')
-        uploadFileAndClickConfirmButton('national_identity_card.pdf')
-        documentUploadConfirmation.verifyUnsuppoertedFileError(documentUploadConfirmationCopy)
-      })
-
-      it('should upload selfie', async () => {
-        goToPassportUploadScreen(`?language=${lang}&async=false&useWebcam=false`)
-        uploadFileAndClickConfirmButton('passport.jpg')
-        uploadFileAndClickConfirmButton('face.jpeg')
-        verificationComplete.verifyUIElements(verificationCompleteCopy)
-        verificationComplete.checkBackArrowIsNotDisplayed()
-      })
-
-      it('should take one selfie using the camera stream', async () => {
-        goToPassportUploadScreen(`?language=${lang}&async=false`)
-        uploadFileAndClickConfirmButton('passport.jpg')
-        documentUploadConfirmation.takeSelfie()
-        documentUploadConfirmation.confirmBtn.click()
-        verificationComplete.verifyUIElements(verificationCompleteCopy)
-        verificationComplete.checkBackArrowIsNotDisplayed()
-      })
-
-      it('should take multiple selfies using the camera stream', async () => {
-        goToPassportUploadScreen(`?language=${lang}&async=false&useMultipleSelfieCapture=true`)
-        uploadFileAndClickConfirmButton('passport.jpg')
-        documentUploadConfirmation.takeSelfie()
-        documentUploadConfirmation.confirmBtn.click()
-        verificationComplete.verifyUIElements(verificationCompleteCopy)
-        verificationComplete.checkBackArrowIsNotDisplayed()
-      })
-
-      it('should return no face found error for selfie', async () => {
-        goToPassportUploadScreen(`?language=${lang}&async=false&useWebcam=false`)
-        uploadFileAndClickConfirmButton('passport.jpg')
-        uploadFileAndClickConfirmButton('llama.jpg')
-        documentUploadConfirmation.verifyNoFaceError(documentUploadConfirmationCopy)
-      })
-
-      it('should return multiple faces error', async () => {
-        goToPassportUploadScreen(`?language=${lang}&async=false&useWebcam=false`)
-        uploadFileAndClickConfirmButton('passport.jpg')
-        uploadFileAndClickConfirmButton('two_faces.jpg')
-        documentUploadConfirmation.verifyMultipleFacesError(documentUploadConfirmationCopy)
-      })
-
-      it('should return glare detected message on front and back of doc', async () => {
-        driver.get(localhostUrl + `?language=${lang}&async=false&useWebcam=false`)
-        welcome.primaryBtn.click()
-        documentSelector.drivingLicenceIcon.click()
-        uploadFileAndClickConfirmButton('identity_card_with_glare.jpg')
-        documentUploadConfirmation.verifyGlareDetectedWarning(documentUploadConfirmationCopy)
-        documentUploadConfirmation.confirmBtn.click()
-        uploadFileAndClickConfirmButton('identity_card_with_glare.jpg')
-        documentUploadConfirmation.verifyGlareDetectedWarning(documentUploadConfirmationCopy)
-      })
-
-      it('should be able to retry document upload', async () => {
-        goToPassportUploadScreen(`?language=${lang}&async=false&useWebcam=false`)
-        documentUpload.getUploadInput()
-        documentUpload.upload('passport.jpg')
-        documentUploadConfirmation.redoBtn.click()
-        uploadFileAndClickConfirmButton('passport.pdf')
-        uploadFileAndClickConfirmButton('face.jpeg')
-        verificationComplete.verifyUIElements(verificationCompleteCopy)
-      })
-
-      it('should be able to submit a document without seeing the document selector screen', async () => {
-        driver.get(localhostUrl + `?language=${lang}&oneDoc=true&async=false&useWebcam=false`)
-        welcome.primaryBtn.click(documentUploadCopy)
-        documentUpload.verifyPassportTitle(documentUploadCopy)
-        uploadFileAndClickConfirmButton('passport.jpg')
-        uploadFileAndClickConfirmButton('face.jpeg')
-        verificationComplete.verifyUIElements(verificationCompleteCopy)
-      })
-    })
+    welcomeScenarios(driver, pageObjects, lang)
+    documentSelectorScenarios(driver, pageObjects, lang)
+    documentScenarios(driver, pageObjects, lang)
+    faceScenarios(driver, pageObjects, lang)
   })
 
   describe('CROSS DEVICE SYNC', async () => {
