+++ test/utils/config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  localhostUrl: 'https://localhost:8080/'
+}
